Refactored how promo action types are handled.
