+++ lib/Action.js
@@ -9,7 +9,7 @@ export default class Action extends Serializable {
   }
 
   static get key() {
-    return Action.name.toLowerCase();
+    return 'action';
   }
 
   static get type() {
