+++ lib/EnhancedEcommerce.js
@@ -1,4 +1,5 @@
 import Action from './Action';
+import ActionPromo from './ActionPromo';
 import Impression from './hits/Impression';
 import Product from './hits/Product';
 import Promo from './hits/Promo';
@@ -76,7 +77,7 @@ export default class EnhancedEcommerce extends Serializable {
     // Checks if action is not of type promo.
     // Promo impressions cannot be sent with promo actions.
     // Reference: https://developers.google.com/analytics/devguides/collection/analyticsjs/enhanced-ecommerce#measuring-promo-impressions
-    if (promos && (action && 'promo' !== action.type)) {
+    if (promos && !(action && ActionPromo.type === action.type)) {
       promos.forEach(promo => {
         query.push(promo.toQueryString());
       });
