+++ lib/hits/Impression.js
@@ -15,7 +15,7 @@ export default class Impression extends Hit {
   }
 
   static get key() {
-    return Impression.name.toLowerCase();
+    return 'impression';
   }
 
   add(batch) {
