+++ lib/hits/Product.js
@@ -16,7 +16,7 @@ export default class Product extends Hit {
   }
 
   static get key() {
-    return Product.name.toLowerCase();
+    return 'product';
   }
 
   add(batch) {
