+++ lib/hits/Promo.js
@@ -11,7 +11,7 @@ export default class Promo extends Hit {
   }
 
   static get key() {
-    return Promo.name.toLowerCase();
+    return 'promo';
   }
 
   add(batch) {
