+++ package.json
@@ -4,7 +4,7 @@
     "type": "git",
     "url": "https://github.com/brandingbrand/react-native-google-analytics.git"
   },
-  "version": "1.11.0",
+  "version": "1.11.1",
   "description": "Google Analytics for React Native!",
   "main": "index.js",
   "author": "Lochlan Wansbrough <lochie@live.com> (http://lwansbrough.com)",
